IG now possible to zoom in and out + move in any direction

+++ src/Model/Sierpinski.java
@@ -106,7 +106,7 @@ public class Sierpinski extends Fractal{
     }
 
     @Override
-    public int coloration(int val) {
+    public int coloration(double val) {
         switch (this.color) {
             case 0 -> {
                 return Color.WHITE.getRGB();
